Add asserts and formatting

+++ include/str.h
@@ -2,7 +2,6 @@
     Immutable strings.
 */
 
-
 #ifndef STR_H_
 #define STR_H_
 
@@ -20,9 +19,9 @@ typedef struct {
 str str_empty(Arena *a);
 str str_new(Arena *a, const char *cstr);
 str str_copy(Arena *a, str str);
-str str_insert(Arena *a, str str, const char ch, uint64_t position);
+str str_insert(Arena *a, str str, const char ch, size_t position);
 str str_concat(Arena *a, str s1, str s2);
-char str_at(str str, uint64_t position);
+char str_at(str str, size_t position);
 bool str_equal(str s1, str s2);
 
 #endif // STR_H_
